services/embedding-worker/requirements.txt
fastapi
uvicorn
pydantic>=2.11.0
pydantic-settings
google-genai
pytest
httpx
